Refactor Repos + LifetimeStats component (data fetching + logic in Repos)
