+++ src/client/components/Pages/Auth/Repos/LifetimeStats/LifetimeStats.tsx
@@ -1,50 +1,19 @@
 import styles from './LifetimeStats.module.css'
 
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
 import LanguageData from './LanguageData'
 import RepoCell from './RepoCell'
 
-import { lifetimeStatsAPI } from '@/client/lib/authAPI'
 import { SHARED_Model__LifetimeStats } from '@/shared/models/apiFields/lifetimeStats'
 import { kbToStr } from '@/shared/utils/toBytesStr'
 
 interface LifetimeStatsProps {
-    accessToken: string
+    lifetimeStats: SHARED_Model__LifetimeStats
 }
 
 export const LifetimeStats: React.FC<LifetimeStatsProps> = (props) => {
-    const { accessToken } = props
-    const [isLoading, setIsLoading] = useState<boolean>(true)
-    const [error, setError] = useState<string>()
-    const [lifetimeStats, setLifetimeStats] = useState<SHARED_Model__LifetimeStats>()
-
-    useEffect(() => {
-        const fetchUserCard = async (accessToken: string) => {
-            setError(undefined)
-            const { success, error, lifetimeStats } = await lifetimeStatsAPI(accessToken)
-            setIsLoading(false)
-            if (!success || lifetimeStats === undefined) {
-                setError(error)
-            } else {
-                setLifetimeStats(lifetimeStats)
-            }
-        }
-
-        fetchUserCard(accessToken)
-    }, [accessToken])
-
-    if (isLoading) {
-        return <div>lifetime stats loading...</div>
-    }
-
-    if (error !== undefined) {
-        return <div>{error}</div>
-    }
-
-    if (lifetimeStats === undefined) {
-        return <div>no lifetime stats found</div>
-    }
+    const { lifetimeStats } = props
 
     const { repos, rc_stats, language_stats } = lifetimeStats
     const { numRepos, numCommits } = rc_stats
